+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - **FEAT**(mouse-actions): Added support for different mouse button actions such as pan, multi-select, and drag-select.
 - **FEAT**(layer-configs): Added `enableKeyboardMultiSelection` and `enableLongPressMultiSelection` to `LayerInteractionConfigs` for dynamically enabling or disabling multi-selection via keyboard or long press.
 - **FEAT**(remove-area): Applied `AnimatedSwitcher` to the remove area for smooth fade-in/out transitions.
+- **FEAT**(crop-rotate-editor): Add `enableFlipAnimation` to `CropRotateEditorConfigs`, which enables flip animation by default.
 
 <br/>